Changes based on feedback
* Extracted LoggerFactory and LoggerFactoryGen
* Moved LoggerName to core and added macro there
* Renamed Platform traits to Compat traits
* Moved Slf4jLoggerFactory into Sl4j4Logger.Factory trait
