+++ core/shared/src/main/scala-2/org/typelevel/log4cats/LoggerNameCompat.scala
@@ -1,0 +1,7 @@
+package org.typelevel.log4cats
+
+import org.typelevel.log4cats.internal.LoggerNameMacro
+
+trait LoggerNameCompat {
+  implicit def name: LoggerName = macro LoggerNameMacro.getLoggerName
+}
