+++ core/shared/src/main/scala/org/typelevel/log4cats/LoggerName.scala
@@ -1,0 +1,5 @@
+package org.typelevel.log4cats
+
+object LoggerName extends LoggerNameCompat
+
+final case class LoggerName(value: String) extends AnyVal
